+++ datapkg/cli/download.py
@@ -1,4 +1,5 @@
 import os
+import fnmatch
 
 from datapkg.cli.base import Command
 from datapkg.download import PackageDownloader
@@ -8,26 +9,43 @@ class DownloadCommand(Command):
     name = 'download'
     summary = 'Download a package'
     min_args = 2
-    max_args = 2
+    max_args = 4
     usage = \
-'''%prog {src-spec} {path}
+'''%prog {src-spec} {path} [format-pattern] [url-pattern]
 
-Install a package specified by {src-spec} to path, e.g.::
+Download a package (i.e. metadata and resources) specified by {src-spec} to
+{path}.
 
+Resources to retrieve are selected based on optional match of resource
+format against glob-style format-pattern and url-pattern, e.g.::
+
+    # download all resources associated with ckan package name
     download ckan://name path-on-disk
 
-Typically this involves downloading the metadata and retrieving all package
-resources.
+    # download only those resources that have format 'csv' (or 'CSV')
+    download ckan://name path-on-disk csv
+
+    # download only those resources that have format starting xml/
+    download ckan://name path-on-disk xml/*
+
+    # download only those resources that have a url starting http://abc (and any format)
+    download ckan://name path-on-disk * http://abc*
 '''
 
     def run(self, options, args):
         spec_from = args[0]
         dest_path_base = args[1]
+        formatpat = args[2] if len(args) > 2 else '*'
+        urlpat = args[3] if len(args) > 3 else '*'
+        def filterfunc(resource, count):
+            match = fnmatch.fnmatch(resource.get('format', ''), formatpat)
+            match = match and fnmatch.fnmatch(resource['url'], urlpat)
+            return match
         index, path = self.index_from_spec(spec_from)
         pkg = index.get(path)
         dest_path = os.path.join(dest_path_base, pkg.name)
-        pkg_downloader = PackageDownloader(verbose=self.verbose)
-        pkg_downloader.download(pkg, dest_path)
+        pkg_downloader = PackageDownloader(verbose=True)
+        pkg_downloader.download(pkg, dest_path, filterfunc)
 
 
 '''